add ci and coverage badges

+++ src/test/java/io/sapient/transmission/NodeDispatcherTest.java
@@ -59,7 +59,7 @@ class NodeDispatcherTest {
                         Registration.StatusDefinition.newBuilder()
                                 .setStatusInterval(
                                         Registration.Duration.newBuilder()
-                                                .setValue(statusReportInterval)
+                                                .setValue((float) statusReportInterval)
                                                 .setUnits(
                                                         Registration.TimeUnits
                                                                 .TIME_UNITS_MILLISECONDS)))
